requirements.txt
# requirements.txt

asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
Werkzeug==3.0.2
